added concern fileable for attached_file
